OASIS Processing Pipeline

This function performs registration to a T1 template using ANTsR transformation

oasis(
  filename,
  skull_strip = TRUE,
  skull_stripfile = NULL,
  correct = TRUE,
  normalize = TRUE,
  normalize_file = NULL,
  retimg = TRUE,
  outfile = NULL,
  template.file = file.path(fsldir(), "data", "standard", "MNI152_T1_1mm_brain.nii.gz"),
  interpolator = "LanczosWindowedSinc",
  other.files = NULL,
  other.outfiles = NULL,
  typeofTransform = "Rigid",
  remove.warp = TRUE,
  outprefix = NULL,
  bet.opts = "-B -f 0.1 -v",
  betcmd = "bet",
  reproducible = TRUE,
  seed = 1,
  ...
)

Arguments

filename
filename of T1 image

skull_strip
do skull stripping with FSL BET

skull_stripfile
Output skull strip filename

correct
do N3 Bias correction

normalize
Normalize data using whitestripe

normalize_file
whitestripe image mask

retimg
return a nifti object from function

outfile
output filename should have .nii or .nii.gz extension

template.file
Filename of template to warp to

interpolator
interpolation done for antsApplyTransforms

other.files
Filenames of other iamges to be transformed with the T1

other.outfiles
Output filenames of other.files to be written

typeofTransform
type of transformed used, passed to antsRegistration

remove.warp
(logical) Should warping images be deleted?

outprefix
Character path of where the warp files should be stored. Required if remove.warp = FALSE

bet.opts
Options passed to fslbet

betcmd
BET command used, passed to fslbet

reproducible
Sets the seed and Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1). See https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues for discussion.

seed
will execute Sys.setenv(ANTS_RANDOM_SEED = seed) before running to attempt a more reproducible result. If NULL, will not set anything, but reproducible must be FALSE.

...
arguments to whitestripe

Value

NULL or object of class nifti for transformed T1 image