Z-score Normalization

This function performs registration to a template using ANTsR and will run zscore_img

reg_zscore(
  t1,
  register = TRUE,
  native = TRUE,
  template.file = file.path(fsldir(), "data", "standard", paste0("MNI152_T1_1mm",
    ifelse(is.null(mask), "", "_brain"), ".nii.gz")),
  typeofTransform = c("Rigid", "Affine"),
  interpolator = "LanczosWindowedSinc",
  t1.outfile = NULL,
  other.files = NULL,
  other.outfiles = NULL,
  mask = NULL,
  mask.outfile = NULL,
  verbose = TRUE,
  reproducible = TRUE,
  seed = 1,
  ...
)

Arguments

t1
filename (or nifti objects) of T1 image

register
Register image to template file

native
If images are registered, should the native space normalized image be returned?

template.file
Filename of template to warp to

typeofTransform
type of transformed used, passed to antsRegistration

interpolator
Interpolation to be performed, passed to antsRegistration

t1.outfile
Output filename of normalized T1 image

other.files
Character filenames or list of nifti objects to normalize. In the same space as T1.

other.outfiles
Character filenames for output normalized files.

mask
File or nifti image of brain mask

mask.outfile
Character filename for output brain mask.

verbose
Print Diagnostic Messages

reproducible
Sets the seed and Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1). See https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues for discussion.

seed
will execute Sys.setenv(ANTS_RANDOM_SEED = seed) before running to attempt a more reproducible result. If NULL, will not set anything, but reproducible must be FALSE.

...
arguments to zscore_img

Value

List of nifti objects or character filenames