Robust Brain Mask from Template

Masks the brain from scan using a rigid body transformation of the template mask to the native brain space and then inflating the mask to adjust for mis-registration.

robust_brain_mask(
  file,
  template.file,
  template.mask = NULL,
  nvoxels = 7,
  ret_mask = FALSE,
  typeofTransform = "Rigid",
  verbose = TRUE,
  ...
)

Arguments

file
File for masking - either filename or class nifti

template.file
Template to warp to original image space

template.mask
Mask of template to use as rough brain mask. If template.file is specified, but template.mask is not, then fslbin(file=template.file) is performed.

nvoxels
Number of voxels in nxnxn direction to inflate, see fslmaths.help for -kernel boxv n -dilM.

ret_mask
Return mask of robust brain

typeofTransform
Transformation for template to image, passed to ants_regwrite.

verbose
Print diagnostic progress

...
Arguments passed to ants_regwrite

Value

Object of class nifti or vector of indices