Create Z-score-to-Template Image

Registers an image to a template, subtracts a mean and sd image, and will return the image either in template space or native.

zscore_template(
  img,
  template.file,
  mean.img,
  sd.img,
  outfile = NULL,
  typeofTransform = "SyN",
  interpolator = "Linear",
  robust = TRUE,
  probs = c(0, 0.999),
  native = TRUE,
  verbose = TRUE,
  ...
)

Arguments

img
Image to be z-scored

template.file
Template file to be registered to. An affine registration will be done usually, so if the image is skull-stripped, use a skull-stripped template and vice versa

mean.img
Mean image to be subtracted

sd.img
SD image to be divded

outfile
Output filename. Will use writenii to write

typeofTransform
Transform for registration. See antsRegistration

interpolator
Interpolator to use, See antsRegistration

robust
Should See robust_window be used on the image

probs
quantiles to constrain the image these define the window sent to robust_window

native
Should a native-space image (default) or template-space image be returned

verbose
Print diagnostic output

...
Options passed to ants_regwrite.

Value

Object of class nifti